Phase 4 P4-5: Xvfb GUI launch tests — 8/8 apps boot (source AND installed)
- New gui_boot_test.py: runs each app's main() in a subprocess under xvfb-run -a;
  a booted app reaches mainloop (killed at timeout = PASS), early exit = FAIL;
  UpdateAuto (non-GUI/network) checked by import + main() signature
- Found + fixed 11 real py3 runtime bugs in the Tk construction path that
  import-smoke could never see: children.values() view (no .extend) x4 incl.
  BasePopup.setFont clones; float-division into %x color formats (Button/Scale/
  Color getIntRgb + 4 sites); range used as list (ScrolledMatrix.refreshSize,
  DangleFrame); range(float) (ViewRamachandranFrame); string.split/capitalize
  (py2) x2; dict_keys.sort (HaddockFrame)
- Fixed installed-state layout bug: 8 gfx/logo dir joins hardcoded the source
  'python/' segment (getTopDirectory()+python) — now getPythonDirectory()
  (packages root in both layouts); apps loaded icons from site-packages/python/...
  which does not exist
- Gates: 8/8 boot source AND installed wheel states; smoke 1646/0/83 and
  1637/0/83, pytest 43/14/10 both states unchanged

+++ ccpnmr2.5/python/ccp/gui/ViewRamachandranFrame.py
@@ -383,7 +383,8 @@ class ViewRamachandranFrame(ScrolledCanvas):
 
         matrix = self.matrix
         canvas = self.canvas
-        nBins = 360 / self.binSize
+        # py3: '/' is float division; range() needs an int (py2 truncated).
+        nBins = int(360 / self.binSize)
         binWidth = self.binWidth = width / float(nBins)
         binHeight = self.binHeight = height / float(nBins)
         # self.cirRadius = min(5, self.binWidth/4.0)
